Create Anagram2.java
This is another implementation of checking if two strings are anagrams : here you can notice that I use "mapping properties" from Mathematics in some way : a mapping is considered to be bijective if it is surjective and injective  at the same time. So I guess I do something like that in this implementation.